requirements.txt
Flask==2.1.3
Flask_Cors==3.0.10
ldap3==2.9.1
pycryptodome==3.16.0